pyproject.toml
[project]
name = "poly-bot"
version = "0.1.0"
description = "Polymarket trading bot with paper trading and pluggable strategies"
requires-python = ">=3.12"
dependencies = [
    # Polymarket official SDK
    "py-clob-client>=0.18.0",
    # HTTP client (async, HTTP/2)
    "httpx>=0.27.0",
    # Data validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # YAML config
    "pyyaml>=6.0.2",
    # Async runtime utilities
    "anyio>=4.4.0",
    # CLI + rich dashboard
    "typer>=0.12.3",
    "rich>=13.7.1",
    # Structured logging
    "structlog>=24.2.0",
    # Async SQLite persistence
    "aiosqlite>=0.20.0",
    # Cryptography for key management
    "cryptography>=42.0.8",
    # Web3 (needed by py-clob-client)
    "web3>=7.0.0",
    "fastapi>=0.135.2",
    "uvicorn[standard]>=0.42.0",
    "websockets>=15.0.1",
    "jinja2>=3.1.6",
]

[project.scripts]
poly-bot = "poly_bot.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["poly_bot"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.2.2",
    "pytest-asyncio>=0.23.7",
    "respx>=0.21.1",
    "mypy>=1.10.0",
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"
